+++ test/e2e/harness/src/index.js
@@ -12,10 +12,13 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { OktaAuth } from '@okta/okta-auth-js';
 import './index.css';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
 import { BrowserRouter as Router } from 'react-router-dom';
+
+const { ISSUER, CLIENT_ID } = process.env;
 
 // To perform end-to-end PKCE flow we must be configured on both ends: when the login is initiated, and on the callback
 // The login page is loaded with a query param. This will select a unique callback url
@@ -25,9 +28,17 @@ const pkce = !!url.searchParams.get('pkce') || url.pathname.indexOf('pkce/callba
 const redirectUri = window.location.origin + (pkce ? '/pkce/callback' : '/implicit/callback');
 const customLogin = !!url.searchParams.get('customLogin');
 
+const oktaAuth = new OktaAuth({
+  issuer: ISSUER,
+  clientId: CLIENT_ID,
+  disableHttpsCheck: true,
+  redirectUri,
+  pkce
+});
+
 ReactDOM.render(
   <Router>
-    <App pkce={pkce} redirectUri={redirectUri} customLogin={customLogin} />
+    <App oktaAuth={oktaAuth} />
   </Router>
   , document.getElementById('root')
 );