+++ test/jest/security.test.js
@@ -6,32 +6,44 @@ import Security from "../../src/Security";
 import { useOktaAuth } from '../../src/OktaContext';
 
 describe('<Security />', () => {
-  const VALID_CONFIG = {
-    issuer: 'https://example.com/oauth2/default',
-    clientId: 'foo',
-    redirectUri: 'https://example.com'
-  };
-  let authService;
+  let oktaAuth;
   let initialAuthState;
   beforeEach(() => {
     initialAuthState = {
       isInitialState: true
     };
-    authService = {
-      on: jest.fn(),
-      updateAuthState: jest.fn(),
-      getAuthState: jest.fn().mockImplementation(() => initialAuthState),
-      _oktaAuth: {
-        token: {
-          isLoginRedirect: jest.fn().mockImplementation(() => false)
-        }
-      }
+    oktaAuth = {
+      userAgent: 'okta/okta-auth-js',
+      options: {},
+      authStateManager: {
+        getAuthState: jest.fn().mockImplementation(() => initialAuthState),
+        subscribe: jest.fn(),
+        updateAuthState: jest.fn(),
+      },
+      isLoginRedirect: jest.fn().mockImplementation(() => false),
     };
   });
 
-  it('gets initial state from authService and exposes it on the context', () => {
+  it('should set userAgent for oktaAuth', () => {
     const mockProps = {
-      authService
+      oktaAuth
+    };
+    mount(<Security {...mockProps} />);
+    expect(oktaAuth.userAgent).toEqual('@okta/okta-react/4.0.0 okta/okta-auth-js');
+  });
+
+  it('should set default restoreOriginalUri callback in oktaAuth.options', () => {
+    oktaAuth.options = {};
+    const mockProps = {
+      oktaAuth
+    };
+    mount(<Security {...mockProps} />);
+    expect(oktaAuth.options.restoreOriginalUri).toBeDefined();
+  });
+
+  it('gets initial state from oktaAuth and exposes it on the context', () => {
+    const mockProps = {
+      oktaAuth
     };
     const MyComponent = jest.fn().mockImplementation(() => {
       const oktaProps = useOktaAuth();
@@ -45,7 +57,7 @@ describe('<Security />', () => {
         </Security>
       </MemoryRouter>
     );
-    expect(authService.getAuthState).toHaveBeenCalled();
+    expect(oktaAuth.authStateManager.getAuthState).toHaveBeenCalled();
     expect(MyComponent).toHaveBeenCalled();
   });
 
@@ -54,16 +66,14 @@ describe('<Security />', () => {
       fromUpdateAuthState: true
     };
     let callback;
-    authService.on.mockImplementation((eventName, fn) => {
-      expect(eventName).toBe('authStateChange');
+    oktaAuth.authStateManager.subscribe.mockImplementation(fn => {
       callback = fn;
     });
-    authService.updateAuthState.mockImplementation(() => {
-      authService.getAuthState.mockImplementation(() => newAuthState);
-      callback();
+    oktaAuth.authStateManager.updateAuthState.mockImplementation(() => {
+      callback(newAuthState);
     });
     const mockProps = {
-      authService
+      oktaAuth
     };
 
     const MyComponent = jest.fn()
@@ -88,22 +98,22 @@ describe('<Security />', () => {
       </MemoryRouter>
     );
 
-    expect(authService.on).toHaveBeenCalledTimes(1);
-    expect(authService.updateAuthState).toHaveBeenCalledTimes(1);
+    expect(oktaAuth.authStateManager.subscribe).toHaveBeenCalledTimes(1);
+    expect(oktaAuth.authStateManager.updateAuthState).toHaveBeenCalledTimes(1);
     expect(MyComponent).toHaveBeenCalledTimes(2);
   });
 
   it('should not call updateAuthState when in login redirect state', () => {
-    authService._oktaAuth.token.isLoginRedirect = jest.fn().mockImplementation(() => true);
+    oktaAuth.isLoginRedirect = jest.fn().mockImplementation(() => true);
     const mockProps = {
-      authService
+      oktaAuth
     };
     mount(
       <MemoryRouter>
         <Security {...mockProps} />
       </MemoryRouter>
     );
-    expect(authService.updateAuthState).not.toHaveBeenCalled();
+    expect(oktaAuth.authStateManager.updateAuthState).not.toHaveBeenCalled();
   });
 
   it('subscribes to "authStateChange" and updates the context', () => {
@@ -118,19 +128,18 @@ describe('<Security />', () => {
     ];
     let callback;
     let stateCount = 0;
-    authService.getAuthState.mockImplementation( () => { 
+    oktaAuth.authStateManager.getAuthState.mockImplementation( () => { 
       return mockAuthStates[stateCount];
     });
-    authService.on.mockImplementation((eventName, fn) => {
-      expect(eventName).toBe('authStateChange');
+    oktaAuth.authStateManager.subscribe.mockImplementation(fn => {
       callback = fn;
     });
-    authService.updateAuthState.mockImplementation(() => {
+    oktaAuth.authStateManager.updateAuthState.mockImplementation(() => {
       stateCount++;
-      callback();
+      callback(mockAuthStates[stateCount]);
     });
     const mockProps = {
-      authService
+      oktaAuth
     };
     const MyComponent = jest.fn()
       // first call
@@ -159,20 +168,20 @@ describe('<Security />', () => {
         </Security>
       </MemoryRouter>
     );
-    expect(authService.on).toHaveBeenCalledTimes(1);
-    expect(authService.updateAuthState).toHaveBeenCalledTimes(1);
+    expect(oktaAuth.authStateManager.subscribe).toHaveBeenCalledTimes(1);
+    expect(oktaAuth.authStateManager.updateAuthState).toHaveBeenCalledTimes(1);
     expect(MyComponent).toHaveBeenCalledTimes(2);
     MyComponent.mockClear();
     act(() => {
       stateCount++;
-      callback();
+      callback(mockAuthStates[stateCount]);
     });
     expect(MyComponent).toHaveBeenCalledTimes(1);
   });
 
   it('should accept a className prop and render a component using the className', () => {
     const mockProps = {
-      authService
+      oktaAuth
     };
     const wrapper = mount(
       <MemoryRouter>
@@ -183,21 +192,74 @@ describe('<Security />', () => {
     expect(wrapper.find(Security).props().className).toBe('foo bar');
   });
 
-  it('Accepts token manager config', () => {
-    const tokenManager = {
-      secure: true,
-      storage: 'cookie'
+  describe('render children', () => {
+    const MyComponent = function() {
+      const { authState } = useOktaAuth();
+      if (authState.isPending) {
+        return <div>loading</div>;
+      }
+
+      if (authState.isAuthenticated) {
+        return <div>Authenticated!</div>;
+      }
+
+      return <div>Not authenticated!</div>;
     };
 
-    const mockProps = Object.assign({}, VALID_CONFIG, {
-      tokenManager,
-      authService
+    it('should render "Authenticated" with preset authState.isAuthenticated as true', () => {
+      initialAuthState = {
+        isAuthenticated: true,
+        isPending: false
+      };
+      const mockProps = {
+        oktaAuth
+      };
+      const wrapper = mount(
+        <Security {...mockProps}>
+          <MyComponent />
+        </Security>
+      );
+      expect(wrapper.find(MyComponent).html()).toBe('<div>Authenticated!</div>');
     });
-    const wrapper = mount(
-      <MemoryRouter>
-        <Security {...mockProps} />
-      </MemoryRouter>
-    );
-    expect(wrapper.find(Security).props().tokenManager).toBe(tokenManager);
+
+    it('should render "Not authenticated" with preset authState.isAuthenticated as false', () => {
+      initialAuthState = {
+        isAuthenticated: false,
+        isPending: false
+      };
+      const mockProps = {
+        oktaAuth
+      };
+      const wrapper = mount(
+        <Security {...mockProps}>
+          <MyComponent />
+        </Security>
+      );
+      expect(wrapper.find(MyComponent).html()).toBe('<div>Not authenticated!</div>');
+    });
+
+    it('should render "loading" with preset authState.isPending as true', () => {
+      initialAuthState = {
+        isPending: true
+      };
+      const mockProps = {
+        oktaAuth
+      };
+      const wrapper = mount(
+        <Security {...mockProps}>
+          <MyComponent />
+        </Security>
+      );
+      expect(wrapper.find(MyComponent).html()).toBe('<div>loading</div>');
+    });
+
+    it('should render error if oktaAuth props is not provided', () => {
+      const wrapper = mount(
+        <Security>
+          <MyComponent />
+        </Security>
+      );
+      expect(wrapper.find(Security).html()).toBe('<p>AuthSdkError: No oktaAuth instance passed to Security Component.</p>');
+    });
   });
 });